+++ packages/sui-studio/bin/sui-studio-init.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-throw new Error(
-  'sui-studio init is not available anymore. Use @s-ui/studio-create instead.'
-)