+++ packages/sui-studio/src/components/style/index.js
@@ -28,5 +28,5 @@ export default function Style({children, id}) {
 }
 
 Style.propTypes = {
-  children: PropTypes.array
+  children: PropTypes.node
 }